# openapi.yaml
{"openapi":"3.0.1","info":{"title":"Filmorate Api","description":"Сервис для базы данных фильмов"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Жанры фильмов","description":"Получение жанров"},{"name":"Пользователи","description":"Управление пользователями"},{"name":"Фильмы","description":"Управление фильмами"},{"name":"Возрастные рейтинги","description":"Получение рейтингов"}],"paths":{"/users":{"get":{"tags":["Пользователи"],"summary":"Получение пользователей","operationId":"getUsers","responses":{"200":{"description":"ОК","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}},"put":{"tags":["Пользователи"],"summary":"Обновление пользователя","operationId":"updateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"ОК","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}},"404":{"description":"Нет пользователя с данным id"}}},"post":{"tags":["Пользователи"],"summary":"Получение пользователя","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"201":{"description":"Пользователь создан","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/users/{id}/friends/{friendId}":{"put":{"tags":["Пользователи"],"summary":"Добавление друга","operationId":"addFriend","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"friendId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"ОК"},"404":{"description":"Нет пользователя с данным id"}}},"delete":{"tags":["Пользователи"],"summary":"Удаление друга","operationId":"deleteFriend","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"friendId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"ОК"},"404":{"description":"Нет пользователя с данным id"}}}},"/films":{"get":{"tags":["Фильмы"],"summary":"Получение фильмов","operationId":"getFilms","responses":{"200":{"description":"ОК","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Film"}}}}}}},"put":{"tags":["Фильмы"],"summary":"Обновление фильма","operationId":"updateFilm","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Film"}}},"required":true},"responses":{"200":{"description":"ОК","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Film"}}}}}},"post":{"tags":["Фильмы"],"summary":"Создание фильма","operationId":"createFilm","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Film"}}},"required":true},"responses":{"201":{"description":"Фильм создан","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Film"}}}}}}},"/films/{id}/like/{userId}":{"put":{"tags":["Фильмы"],"summary":"Добавление лайка фильму","operationId":"addLike","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"ОК"},"404":{"description":"Нет фильма с данным id"}}},"delete":{"tags":["Фильмы"],"summary":"Удаление лайка","operationId":"deleteLike","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"ОК"},"404":{"description":"Нет фильма с данным id"}}}},"/users/{id}":{"get":{"tags":["Пользователи"],"summary":"Получение пользователя","operationId":"getUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"ОК","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}},"404":{"description":"Нет пользователя с данным id"}}}},"/users/{id}/friends":{"get":{"tags":["Пользователи"],"summary":"Получение друзей пользователя","operationId":"getFriends","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"ОК","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}},"404":{"description":"Нет пользователя с данным id"}}}},"/users/{id}/friends/common/{otherId}":{"get":{"tags":["Пользователи"],"summary":"Получение общих друзей двух пользователей","operationId":"getCommonFriends","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"otherId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"ОК","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}},"404":{"description":"Нет пользователя с данным id"}}}},"/mpa":{"get":{"tags":["Возрастные рейтинги"],"summary":"Получение рейтингов","operationId":"getRatings","responses":{"200":{"description":"ОК","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Mpa"}}}}}}}},"/mpa/{id}":{"get":{"tags":["Возрастные рейтинги"],"summary":"Получение рейтинга","operationId":"getRating","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"ОК","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Mpa"}}}},"404":{"description":"Нет рейтинга с данным id"}}}},"/genres":{"get":{"tags":["Жанры фильмов"],"summary":"Получение жанров","operationId":"getGenres","responses":{"200":{"description":"ОК","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Genre"}}}}}}}},"/genres/{id}":{"get":{"tags":["Жанры фильмов"],"summary":"Получение жанра","operationId":"getGenre","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"ОК","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Genre"}}}},"404":{"description":"Нет жанра с данным id"}}}},"/films/{id}":{"get":{"tags":["Фильмы"],"summary":"Получение фильма","operationId":"getFilm","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"ОК","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Film"}}}},"404":{"description":"Нет фильма с данным id"}}}},"/films/popular":{"get":{"tags":["Фильмы"],"summary":"Получение популярных фильмов","operationId":"getPopular","parameters":[{"name":"count","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"ОК","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Film"}}}}}}}}},"components":{"schemas":{"User":{"type":"object","properties":{"id":{"type":"integer","description":"Id пользователя","format":"int32","example":1},"email":{"type":"string","description":"Почта пользователя","example":"vasya@mail.com"},"login":{"type":"string","description":"Логин пользователя","example":"vasya123"},"name":{"type":"string","description":"Имя пользователя","example":"Вася"},"birthday":{"type":"string","description":"Дата рождения","example":"2000-01-01T00:00:00"},"friends":{"uniqueItems":true,"type":"array","description":"Список id друзей","example":[1,2,3],"items":{"type":"integer","description":"Список id друзей","format":"int32"}}},"description":"Сущность пользователя"},"Film":{"type":"object","properties":{"id":{"type":"integer","description":"Id фильма","format":"int32","example":1},"name":{"type":"string","description":"Название фильма","example":"Отступники"},"description":{"type":"string","description":"Описание фильма","example":"Лучшее описание"},"releaseDate":{"type":"string","description":"Дата выхода","example":"2000-01-01T00:00:00"},"duration":{"type":"integer","description":"Продолжительность в минутах","format":"int32","example":90},"mpa":{"$ref":"#/components/schemas/Mpa"},"likes":{"uniqueItems":true,"type":"array","description":"Id пользователей, поставивших лайк","example":[3,1,2],"items":{"type":"integer","description":"Id пользователей, поставивших лайк","format":"int32"}},"genres":{"type":"array","description":"Жанры фильма","example":"[Драма,Боевик]","items":{"$ref":"#/components/schemas/Genre"}}},"description":"Сущность фильма"},"Genre":{"type":"object","properties":{"id":{"type":"integer","description":"Id жанра","format":"int32","example":1},"name":{"type":"string","description":"Название жанра","example":"Комедия"}},"description":"Сущность жанра","example":"[Драма,Боевик]"},"Mpa":{"type":"object","properties":{"id":{"type":"integer","description":"Id рейтинга","format":"int32","example":1},"name":{"type":"string","description":"Название рейтинга","example":"PG-13"}},"description":"Сущность возрастного рейтинга","example":"R"}}}}